Refactor code + make model files smaller + optimize warmup
