+++ package.json
@@ -15,7 +15,9 @@
     "@popperjs/core": "^2.11.8",
     "@react-hook/window-size": "^3.1.1",
     "@reduxjs/toolkit": "^1.9.7",
-    "@tensorflow/tfjs": "^4.14.0",
+    "@tensorflow/tfjs-backend-webgl": "^4.15.0",
+    "@tensorflow/tfjs-converter": "^4.15.0",
+    "@tensorflow/tfjs-core": "^4.15.0",
     "bootstrap": "^5.3.0",
     "chess.js": "^1.0.0-beta.6",
     "delaunator": "^5.0.0",
@@ -39,6 +41,7 @@
     "eslint": "^8.49.0",
     "eslint-plugin-react": "^7.33.2",
     "gh-pages": "^5.0.0",
-    "vite": "^3.2.3"
+    "vite": "^3.2.3",
+    "vite-bundle-visualizer": "^1.0.0"
   }
 }
