+++ src/components/common/icon.jsx
@@ -1,0 +1,7 @@
+const Icon = ({ iconName }) => {
+  return (
+    <i className={`h4 bi ${iconName}`} />
+  )
+}
+
+export default Icon;
