+++ src/components/common/index.jsx
@@ -1,0 +1,13 @@
+import SidebarButton from "./sidebarButton";
+import Icon from "./icon";
+import Corners from "./corners";
+import CornersButton from "./cornersButton";
+import Display from "./display";
+import HomeButton from "./homeButton";
+import Marker from "./marker";
+import PgnButton from "./pgnButton";
+import Sidebar from "./sidebar";
+import Container from "./container";
+
+export { Sidebar, SidebarButton, Icon, Corners, CornersButton, 
+  Display, HomeButton, Marker, PgnButton, Container };
