+++ src/components/common/sidebarButton.jsx
@@ -1,0 +1,9 @@
+const SidebarButton = (props) => {
+  return (
+    <button onClick={props.onClick} className="btn btn-dark btn-sm btn-outline-light w-100">
+      {props.children}
+    </button>
+  )
+}
+
+export default SidebarButton;
