+++ src/components/export/buttons/index.jsx
@@ -1,0 +1,4 @@
+import ExportButton from "./exportButton";
+import StudyButton from "./studyButton";
+
+export { ExportButton, StudyButton };
