+++ src/components/record/buttons/index.jsx
@@ -1,0 +1,4 @@
+import HelpButton from "./helpButton";
+import RecordButton from "./recordButton";
+
+export { HelpButton, RecordButton };
