+++ src/components/upload/buttons/index.jsx
@@ -1,0 +1,5 @@
+import PlayButton from "./playButton";
+import RestartButton from "./restartButton";
+import VideoButton from "./videoButton";
+
+export { PlayButton, RestartButton, VideoButton };
