+++ src/components/upload/buttons/playButton.jsx
@@ -1,0 +1,19 @@
+import { SidebarButton, Icon } from "../../common";
+
+const PlayButton = ({ videoRef, playing, setPlaying }) => {
+  const handleClick = (e) => {
+    e.preventDefault();
+    
+    if (videoRef.current.src.startsWith("blob")) {
+      setPlaying(!playing);
+    }
+  }
+
+ return (
+    <SidebarButton onClick={handleClick}>
+      <Icon iconName={playing ? "bi-pause" : "bi-play"} />
+    </SidebarButton>
+  );
+};
+
+export default PlayButton;
