+++ src/components/upload/upload.jsx
@@ -4,6 +4,7 @@ import VideoSidebar from "./videoSidebar.jsx";
 import { useOutletContext } from "react-router-dom";
 import { useDispatch } from 'react-redux';
 import { cornersReset } from '../../slices/cornersSlice.jsx';
+import { Container } from "../common";
 
 const Upload = () => {
   const [piecesModelRef, xcornersModelRef] = useOutletContext();
@@ -26,13 +27,13 @@ const Upload = () => {
   }, []);
 
   return (
-    <div className="d-flex bg-dark h-100">
+    <Container>
       <VideoSidebar videoRef={videoRef} piecesModelRef={piecesModelRef} xcornersModelRef={xcornersModelRef} 
       canvasRef={canvasRef} setText={setText} sidebarRef={sidebarRef} playing={playing} setPlaying={setPlaying}
       text={text} />
       <Video modelRef={piecesModelRef} videoRef={videoRef} canvasRef={canvasRef} sidebarRef={sidebarRef} 
       playing={playing} setPlaying={setPlaying} playingRef={playingRef} setText={setText} />
-    </div>
+    </Container>
   );
 };
