+++ src/utils/render/renderBox.jsx
@@ -1,5 +1,5 @@
 import * as Constants from "../constants.jsx";
-import * as tf from "@tensorflow/tfjs";
+import * as tf from "@tensorflow/tfjs-core";
 import { setupCtx, drawBox, drawPoints, drawPolygon } from "./common.jsx";
 
 export const renderBoxes = (canvasRef, boxes, scores, centers, boundary, squares) => {
@@ -9,7 +9,7 @@ export const renderBoxes = (canvasRef, boxes, scores, centers, boundary, squares
     let maxConf = Array(64).fill(0.5);
     let idx = Array(64).fill(-1);
 
-    const confArray = scores.max(1).arraySync();
+    const confArray = tf.max(scores, 1).arraySync();
     for (let i = 0; i < squares.length; i++) {
       const conf = confArray[i];
       const square = squares[i];
@@ -26,10 +26,10 @@ export const renderBoxes = (canvasRef, boxes, scores, centers, boundary, squares
 
     idx = idx.filter(i => i != -1);
     const squaresTensor = tf.tensor1d(squares);
-    const bboxConfClsSquare = tf.concat([boxes.gather(idx),
-    scores.max(1).gather(idx).expandDims(1),
-    scores.argMax(1).gather(idx).expandDims(1),
-    squaresTensor.gather(idx).expandDims(1)],
+    const bboxConfClsSquare = tf.concat([tf.gather(boxes, idx),
+    tf.expandDims(tf.gather(tf.max(scores, 1), idx), 1),
+    tf.expandDims(tf.gather(tf.argMax(scores, 1), idx), 1),
+    tf.expandDims(tf.gather(squaresTensor, idx), 1)],
     1);
     return bboxConfClsSquare.arraySync();
   });
