+++ vite.config.js
@@ -3,7 +3,9 @@ import react from "@vitejs/plugin-react";
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [react()],
+  plugins: [
+    react()
+  ],
   build: {
     chunkSizeWarningLimit: 2000, // handle warning on vendor.js bundle size
   },
